requirements.txt
flask
flask_cors
google-cloud-aiplatform
python-dotenv
gunicorn
